refactor: split shop command
