+++ slash_commands/subcommands/shop/index.js
@@ -1,0 +1,11 @@
+const { custom } = require("./custom");
+const { jeux } = require("./jeux");
+const { list } = require("./list");
+const { sell } = require("./sell")
+
+module.exports = {
+    custom,
+    jeux,
+    list,
+    sell,
+}
